refactor: move PasswordRuleSerializer to web/serializers.py
