+++ src/bk-user/bkuser/apis/web/personal_center/serializers.py
@@ -24,6 +24,7 @@ from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 
 from bkuser.apis.web.personal_center.constants import PhoneOrEmailUpdateRestrictionEnum
+from bkuser.apis.web.serializers import PasswordRuleSerializer
 from bkuser.apis.web.tenant_setting.serializers import BuiltinFieldOutputSLZ
 from bkuser.apps.data_source.models import (
     DataSourceDepartmentUserRelation,
@@ -36,7 +37,6 @@ from bkuser.biz.validators import validate_logo, validate_user_extras, validate_
 from bkuser.common.constants import TIME_ZONE_CHOICES, BkLanguageEnum
 from bkuser.common.desensitize import desensitize_email, desensitize_phone
 from bkuser.common.hashers import check_password
-from bkuser.common.serializers import PasswordRuleSerializer
 from bkuser.common.validators import validate_phone_with_country_code
